Repair to read modules.
